notification_handler/requirements.txt
mysql-connector==2.2.9
kafka-python==2.0.2
python-dotenv==1.0.1
